dds poc working with low fps
